PLEASE FINAL COMMIT

+++ database/migrations/2016_11_19_130615_create_event_table.php
@@ -16,7 +16,7 @@ class CreateEventTable extends Migration
         Schema::create('event', function (Blueprint $table) {
 
             $table->increments('id');
-            $table->integer('preis');
+            $table->double('preis');
             $table->text('kurzbeschrieb');
             $table->text('tag1')->nullable();
             $table->text('tag2')->nullable();
